refactor(kernel): consolidate grouped FP8 module
